add ejs, sass features
