+++ src/pages/home/header.ejs
@@ -1,0 +1,2 @@
+<header><%= data.title %></header>
+<p><%= data.desc %></p>
